# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: OutreachHub API
  description: |
    OutreachHub 智能拓客与邮件营销平台 API。

    ## 认证
    所有 API 请求需要通过 `Authorization: Bearer <token>` header 或 `auth-token` cookie 进行认证。

    ## 通用响应格式
    - 成功：`{ "success": true, "data": ... }`
    - 错误：`{ "success": false, "error": { "code": "...", "message": "..." } }`
    - 分页：`{ "success": true, "data": [...], "pagination": { "page": 1, "limit": 20, "total": 100, "pages": 5 } }`
  version: 1.0.0
  contact:
    name: OutreachHub Support
    email: support@outreachhub.com
servers:
  - url: http://localhost:3030
    description: 本地开发
  - url: https://app.outreachhub.com
    description: 生产环境
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: 'API Key (格式: oh_xxx)'
  responses:
    RateLimitUnavailable:
      description: API Key 分布式限流服务暂不可用
      headers:
        Retry-After:
          description: 建议重试等待秒数
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer
    Contact:
      type: object
      properties:
        id:
          type: string
        fullName:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        title:
          type: string
        companyId:
          type: string
        status:
          type: string
          enum: [NEW, CONTACTED, INTERESTED, NOT_INTERESTED, QUALIFIED, CONVERTED, UNREACHABLE]
        tags:
          type: array
          items:
            type: string
        emails:
          type: array
          items:
            $ref: '#/components/schemas/ContactEmail'
        createdAt:
          type: string
          format: date-time
    ContactEmail:
      type: object
      properties:
        id:
          type: string
        address:
          type: string
        type:
          type: string
          enum: [PROFESSIONAL, PERSONAL]
        isVerified:
          type: boolean
...
paths:
  /api/contacts:
    get:
      summary: 获取联系人列表
      tags: [联系人]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          description: 搜索姓名或邮箱
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [NEW, CONTACTED, INTERESTED, NOT_INTERESTED, QUALIFIED, CONVERTED, UNREACHABLE]
        - name: pool
          in: query
          description: 公海/私海筛选
          schema:
            type: string
            enum: [PUBLIC, PRIVATE]
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Contact'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
    post:
      summary: 创建联系人
      tags: [联系人]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              anyOf:
                - required: [fullName]
                - required: [firstName]
              properties:
                fullName:
                  type: string
                firstName:
                  type: string
                lastName:
                  type: string
                title:
                  type: string
                companyId:
                  type: string
                emails:
                  type: array
                  items:
                    type: string
                tags:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: 创建成功
        '403':
          description: 权限不足或限额已满
...
  /api/v1/contacts:
    get:
      summary: 获取联系人列表（公开 API v1）
      description: |
        需要 `contacts:read` 或 `contacts:manage` 权限。
        所有查询自动过滤为当前租户数据，无法跨租户访问。
      tags: [公开 API v1]
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: page
          in: query
          description: 页码（从 1 开始）
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          description: 每页数量（最大 100）
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: search
          in: query
          description: 搜索姓名或邮箱
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [NEW, CONTACTED, INTERESTED, NOT_INTERESTED, QUALIFIED, CONVERTED, UNREACHABLE]
      responses:
        '200':
          description: 成功
          headers:
            X-RateLimit-Limit:
              description: 当前时间窗口最大请求数
              schema:
                type: integer
            X-RateLimit-Remaining:
              description: 当前窗口剩余请求数
              schema:
                type: integer
            X-RateLimit-Reset:
              description: 窗口重置的 Unix 时间戳（秒）
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Contact'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          description: 未认证或 API Key 无效/过期/已撤销
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足或缺少租户关联
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: 请求频率超出限制
          headers:
            Retry-After:
              description: 建议等待秒数
              schema:
                type: integer
            X-RateLimit-Limit:
              schema:
                type: integer
            X-RateLimit-Remaining:
              schema:
                type: integer
              example: 0
            X-RateLimit-Reset:
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        example: RATE_LIMIT_EXCEEDED
                      message:
                        type: string
                  retryAfter:
                    type: integer
        '503':
          $ref: '#/components/responses/RateLimitUnavailable'
    post:
      summary: 创建联系人（公开 API v1）
      description: |
        需要 `contacts:write` 或 `contacts:manage` 权限。
        受套餐联系人数量限制。
      tags: [公开 API v1]
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              anyOf:
                - required: [fullName]
                - required: [firstName]
              properties:
                fullName:
                  type: string
                  description: 全名（与 firstName 二选一）
                firstName:
                  type: string
                lastName:
                  type: string
                title:
                  type: string
                companyId:
                  type: string
                emails:
                  type: array
                  items:
                    type: string
                tags:
                  type: array
                  items:
                    type: string
                notes:
                  type: string
      responses:
        '201':
          description: 创建成功
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
            X-RateLimit-Remaining:
              schema:
                type: integer
            X-RateLimit-Reset:
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Contact'
        '400':
          description: 参数校验失败
        '401':
          description: 未认证或 API Key 无效
        '403':
          description: 权限不足或限额已满
        '429':
          description: 请求频率超出限制
          headers:
            Retry-After:
              schema:
                type: integer
        '503':
          $ref: '#/components/responses/RateLimitUnavailable'
...